some modify code

+++ src/app/module/Blog Model/blog.Controller.ts
@@ -3,11 +3,20 @@ import { blogServices } from "./blog.Services"
 import { StatusCodes } from "http-status-codes"
 import sendResponse from "../../utils/sendRespose"
 import asyncFunc from "../../utils/asyncFunc"
+import { blogInterface } from "./blog.Interface"
+import AppError from "../../errors/AppError"
+import { blogModel } from "./blog.Model.Schema"
 
 // create a blog 
 const createBlog = asyncFunc(async (req: Request, res: Response) => {
-    const body = req.body
-    const result = await blogServices.createBlogIntoDb(body)
+    if(!req?.user){
+        throw new AppError (StatusCodes.BAD_GATEWAY,"You must be logged in to create a blog.")
+    }
+    const { title, content } = req.body
+
+    const payload: blogInterface = { title, content, author: req?.user.id, isPublished: true }
+
+    const result = await blogServices.createBlogIntoDb(payload)
     sendResponse(res, {
         statusCode: StatusCodes.CREATED,
         message: 'Blog created successfully',
@@ -27,7 +36,9 @@ const getAllBlog = asyncFunc(async (req: Request, res: Response) => {
 
 // get single blogs 
 const getSingleBlog = asyncFunc(async (req: Request, res: Response) => {
+
     const { id } = req.params
+    
     const result = await blogServices.getSingleBlogDb(id)
     sendResponse(res, {
         statusCode: StatusCodes.OK,
@@ -39,6 +50,18 @@ const getSingleBlog = asyncFunc(async (req: Request, res: Response) => {
 // update a blog 
 const updateSingleBlog = asyncFunc(async (req: Request, res: Response) => {
     const { id } = req.params
+    const userId = req?.user?.id
+    
+    const blog = await blogModel.findOne({_id: id})
+
+    if(!blog){
+        throw new AppError(StatusCodes.NOT_FOUND, 'Blog not found')
+    }
+
+    if(blog.author.toString() !== userId){
+        throw new AppError(StatusCodes.FORBIDDEN, "You are not authorized to update this blog")
+    }
+
     const payload = req.body
     const result = await blogServices.updateSingleBlogDb(id, payload)
     sendResponse(res, {
@@ -51,6 +74,18 @@ const updateSingleBlog = asyncFunc(async (req: Request, res: Response) => {
 // delete a blog  
 const deleteSingleBlog = asyncFunc(async (req: Request, res: Response) => {
     const { id } = req.params
+    const userId = req?.user?.id
+    
+    const blog = await blogModel.findOne({_id: id})
+
+    if(!blog){
+        throw new AppError(StatusCodes.NOT_FOUND, 'Blog not found')
+    }
+
+    if(blog.author.toString() !== userId){
+        throw new AppError(StatusCodes.FORBIDDEN, "You are not authorized to delete this blog")
+    }
+
     await blogServices.deleteSingleBlogDb(id)
     sendResponse(res, {
         statusCode: StatusCodes.OK,
